+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Fixed
+- MarketHistory typings which is a list by default.
+
 ## [4.0.0] - 2019-07-09
 ### BREAKING CHANGES
 - Renamed constructESIURL to constructESIUrl.